+++ pkg/kubelet/certificate/BUILD
@@ -24,8 +24,8 @@ go_library(
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes/typed/certificates/v1beta1:go_default_library",
         "//staging/src/k8s.io/client-go/rest:go_default_library",
+        "//staging/src/k8s.io/client-go/rest/dialer:go_default_library",
         "//staging/src/k8s.io/client-go/util/certificate:go_default_library",
-        "//staging/src/k8s.io/client-go/util/connrotation:go_default_library",
         "//vendor/github.com/prometheus/client_golang/prometheus:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
     ],
