+++ staging/src/k8s.io/client-go/rest/dialer/dialer_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package connrotation
+package dialer
 
 import (
 	"context"
